libnetwork: move all SandboxOptions to a separate file
Just a minor reorganisation; move options together into a dedicated file.
